game_server/requirements.txt
flask==2.3.2
black==23.7.0
pytest==7.4.0
requests==2.31.0
urllib3==1.26.6  # Don't change version.
python-dotenv==1.0.0
gunicorn==21.1.0
pymongo==4.4.1
jsonschema~=4.18.4